refactor(FE-11): refactored  useMapWidget and divided into several separate smaller hooks
- refactored App component according to ES Lint rules.
- added consts for MapWidget
