+++ src/app/App.tsx
@@ -1,12 +1,12 @@
 import { Providers } from './providers';
 import { AppRouter } from './router';
 
-function App() {
+const App = () => {
 	return (
 		<Providers>
 			<AppRouter />
 		</Providers>
 	);
-}
+};
 
 export default App;
