+++ src/widgets/map-widget/config/consts.ts
@@ -1,0 +1,4 @@
+import { LatLngExpression } from 'leaflet';
+
+export const MAP_INITIAL_CENTER: LatLngExpression = [51.505, 0];
+export const MAP_INITIAL_ZOOM: number = 3;
