+++ src/widgets/map-widget/config/index.ts
@@ -1,0 +1,1 @@
+export { MAP_INITIAL_CENTER, MAP_INITIAL_ZOOM } from './consts';
