+++ src/widgets/map-widget/model/types.ts
@@ -1,15 +1,7 @@
-import { DivIcon, LatLngExpression, Map } from 'leaflet';
+import { DivIcon } from 'leaflet';
 
 import { Stop } from 'entities/stop';
 import { Trip } from 'entities/trip';
-
-export interface IMapViewModel {
-	initialCenter: LatLngExpression;
-	initialZoom: number;
-	markers: MapMarkerData[];
-	routes: MapRouteData[];
-	onMapLoad: (mapInstance: Map) => void;
-}
 
 export interface MapRouteData {
 	readonly key: string;
